requirements.txt
Pillow==9.0.0
Flask==1.1.2
pytesseract~=0.3.8